requirements.txt
ffmpeg-python
openai
jiwer
python-dotenv
